feat: add frame tick interpolation for slow motion
